Python/TOP 10 SONGS OF THE WEEK/requirement.txt
request
bs4
lxml
